Updated the proofs enum
