+++ crates/errors/src/lib.rs
@@ -124,16 +124,28 @@ pub enum CommitmentError {
     CurrentCommitmentMismatch,
 }
 
-#[derive(Error, Debug)]
+#[derive(Error, Clone, Debug)]
 pub enum ProofError {
-    #[error("generating proof: {0}")]
-    GenerationError(String),
-    #[error("verifying proof: {0}")]
+    #[error("service proof is missing from batch for create account verification: {0}")]
+    MissingServiceProof(String),
+    #[error("service challenge is missing for create account verification: {0}")]
+    MissingServiceChallenge(String),
+    #[error("encoding error: {0}")]
+    EncodingError(String),
+    #[error("account update error: {0}")]
+    AccountError(String),
+    #[error("proof verification error: {0}")]
     VerificationError(String),
-    #[error("deserializing G1Affine point")]
-    G1AffineDeserializationError,
-    #[error("unpacking proof components: {0}")]
-    ProofUnpackError(String),
-    #[error("invalid proof format")]
-    InvalidFormatError,
+    #[error("existence error: {0}")]
+    ExistenceError(String),
+    #[error("nonexistence error: {0}")]
+    NonexistenceError(String),
+    #[error("Transaction error: {0}")]
+    TransactionError(String),
+}
+
+impl From<bincode::Error> for ProofError {
+    fn from(err: bincode::Error) -> Self {
+        ProofError::EncodingError(err.to_string())
+    }
 }
