+++ verification_keys/keys.json
@@ -1,1 +1,1 @@
-{"base_vk": "0x0057419480a88f5392354ca72cda440060507d726258793ab4cbd56455584722", "recursive_vk": "0x00eb84c81095d76d5e81c5f469980176bccedfb349bb1034e7cd1c8307d9af90"}
+{"base_vk": "0x00590b22eb65ab28d63742754a82fa1c9c868749c129e4edb7fa6e6e2022d099", "recursive_vk": "0x00d4431e19dd1a83d2fdbcb44ce8aa7de4953cbe6ebd9d5bf5e6c2e83a745765"}
